chore(refactor): Extract inner classes, abstract common test pieces and fix javadoc
