+++ config/impl-base/src/main/java/org/jboss/arquillian/config/impl/extension/PropertyResolver.java
@@ -1,0 +1,5 @@
+package org.jboss.arquillian.config.impl.extension;
+
+interface PropertyResolver {
+    String getValue(String key);
+}
